stock_subscriptions/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multitasking==0.0.9
mysql-connector-python==8.0.20
numpy==1.18.4
pandas==1.0.3
protobuf==3.11.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
yfinance==0.1.54
